requirements.txt
# Core dependencies
fastapi[standard]>=0.112.0,<0.113.0
pydantic>=2.7.0,<3.0.0
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
httpx==0.27.2
tenacity==9.0.0
python-dateutil==2.9.0.post0
feedparser==6.0.11

# RabbitMQ
aio-pika==9.3.0

# HTTP client for API calls
aiohttp==3.9.1

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Web scraping
beautifulsoup4==4.12.2
lxml==4.9.3

# AI/ML (for query refinement)
openai==1.3.7

# File handling
aiofiles==23.2.1

# Data processing
pandas==2.1.4
numpy==1.25.2

# B2 Storage
b2sdk==1.24.0

# Logging and monitoring
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0